Rename project to GitLab.NET
Using GitLabAPI.NET was causing ambiguity problems between the
GitLabAPI class and the GitLabAPI namespace. To fix this, we renamed
the namespace to GitLab.NET because it is a much nicer name to call the
project. Using GitLabAPI for the root class is acceptable, because it
is more descriptive of what the class' purpose is.