Modify design for better implementation
GitLabAPI is now an abstract class. This decision was made because I
feel it is better to use a more modular design than to have consumers
simply create an instance of GitLabAPI and use that for everything.
Doing things this way is more in line with the way REST works, and also
makes for much cleaner code, with better separation of concerns.

The tests for GitLabAPI had to be modified to create a class that
inherits from GitLabAPI in order to keep the tests. We still want the
tests because we want to make sure that GitLabAPI's implementation is
tested, rather than testing every class that inherits from GitLabAPI.