Rename GitLabAPI to GitLabBase
GitLabBase is a better name for this abstract class. This will allow us
to create a new GitLabAPI class which serves as a class with instances
of all the subclasses. This allows more simplicity for consumers,
because they can use methods like GitLabAPI.Users.Search(). Consumers
will still be able to create their own instances of GitLabBase
subclasses if they prefer. This makes the API more versatile, and
allows it to suit more use cases.