Add ConfigureAwait.Fody and do some cleanup
Async methods can block the UI thread. It's recommended to add
ConfigureAwait(false) to all of your await calls in a library to
prevent this as well as improve performance. That uglifies your code to
death though. Installing ConfigureAwait.Fody is another option that
rewrites the IL on compile to add this, without uglifying your source
code.

Also did some renaming, modifying of comments, and fixing of
accessibility and constructors.