Refactor and clean up core classes and tests
Changed scope on some classes.

No longer allowing test project to have access to internal classes in
the main project. This is because we shouldn't have to test the
internal implementation details of the main project, but rather the
tests on the public classes should cover the internal implementation
enough. That, and testing implementation details creates brittle tests
which are bad.

Also cleaned up a few files, specifically moved HeaderHelper's
functionality directly into PaginatedResult rather than using an
extension method, as it is only used within PaginatedResult anyway, so
doesn't warrant an extension method.