Change all types to nullable
- Also changed any int types to uint that warranted it. uint is better
to use for id fields, counts, etc as we aren't expecting negative
values here.
- We use nullable types for ResponseModels because the API may omit
values in its response. If the API omits a value that can't be null, we
get default values such as '0', 'false', etc. This can lead to
unexpected results, and as such it is better to use a nullable value so
it is clear whether the result was returned or not.